changed gain control

+++ Basis/Packages/com.avionblock.opussharp/OpusSharp.Core/Extensions/OpusDecoderExtensions.cs
@@ -15,7 +15,7 @@ namespace OpusSharp.Core.Extensions
         /// <param name="gain">The gain to set.</param>
         /// <exception cref="OpusException" />
         /// <exception cref="ObjectDisposedException" />
-        public static void SetGain(this OpusDecoder decoder, short gain)
+        public static void SetGain(this OpusDecoder decoder, int gain)
         {
             decoder.Ctl(DecoderCTL.OPUS_SET_GAIN, gain);
         }
